Changes on 4 june

+++ lib/features/industry/register_truck_movements/widgets/in_preminar_info_widget.dart
@@ -26,7 +26,7 @@ class InPreliminarInfoWidget extends StatelessWidget {
             SizedBox(height: 28.h,),
             CustomTile(
                 title: "Número de guía",
-                subText: viajesModel.guideNumber.toString()
+                subText: viajesModel.guideNumber.toInt().toString()
             ),
             CustomTile(
                 title: "Placa",
